Add files via upload
This is a code that draws Mandelbrot sets, a type of strange object with recursive edges that are self-similar at any level of zoom. The basic code for this comes from (credit: https://docs.scipy.org/doc/numpy/user/quickstart.html), with modifications around iterations, placement, and color scheme. Enjoy playing around with it.